MGMT-17019: Separate CLI commands and move to internal/cli package
